+++ src/base/shape.h
@@ -5,7 +5,7 @@
 class Material;
 
 struct Intersection {
-        float t;
+        double t;
         Point p;
         Vector3 n;
         const Material *material_ptr;
@@ -15,8 +15,8 @@ class Shape {
     public:
         __device__ virtual ~Shape() {}
 
-        __device__ virtual bool intersect(Intersection &intersection, const Ray &ray, float t_min,
-                                          float t_max) const = 0;
+        __device__ virtual bool intersect(Intersection &intersection, const Ray &ray, double t_min,
+                                          double t_max) const = 0;
 
         __device__ virtual const Material *get_material_ptr() const = 0;
 };
